Update Dashboard: update Manage Customers
