+++ SavingsBook/EditSavingBook.resx
@@ -117,4 +117,15 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
+  <data name="btnTypeMenu.Image" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
+    <value>
+        iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAO
+        vAAADrwBlbxySQAAABl0RVh0U29mdHdhcmUAd3d3Lmlua3NjYXBlLm9yZ5vuPBoAAACySURBVDhPYxgF
+        EOCw5EuFw+Iv/0nCQD1Q7QwM1gs+JQPxfxJxMlQ7BJjP/dBjPufDf6Lw7A9dUG1I4P9/RsPpbxcYzXj3
+        Hx82nvFuGUPDfyaoLlRgPPM/q960Nzv0pr75jxVPebNXZdJtdqhy7MB45jMuzYmvjgPxfzR8RnvqKx6o
+        MvxAreeZiErf85uqfS/+Q/FdxQkvxaHSxAHZrpfKCl3P7sh3P78NYkOFRwEGYGAAALVKy5YhlXkVAAAA
+        AElFTkSuQmCC
+</value>
+  </data>
 </root>
