+++ SavingsBook/LoginForm.cs
@@ -15,6 +15,10 @@ namespace GUI
         public LoginForm()
         {
             InitializeComponent();
+            // Thiết lập không cho phép thay đổi kích thước Form
+            this.FormBorderStyle = FormBorderStyle.FixedSingle;
         }
+
+
     }
 }
