+++ SavingsBook/LoginForm.resx
@@ -121,7 +121,7 @@
   <data name="$this.BackgroundImage" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
         iVBORw0KGgoAAAANSUhEUgAAA4QAAAJYCAIAAAC1p7+MAAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAO
-        wAAADsABataJCQAA3FNJREFUeF7t/OeTpvlxnQnv3zhd3a19V7sr7Wq1KxEzI0qiSEkURUoUKToQBA0G
+        vgAADr4B6kKxwAAA3FNJREFUeF7t/OeTpvlxnQnv3zhd3a19V7sr7Wq1KxEzI0qiSEkURUoUKToQBA0G
         hCdAYOC99977gffee0uQIN88edLd5umq6qpqDp45GSdO3DWfeyKuuH755P/y9zdv/sPNf7Lsf/IPN//X
         n9+wj//15zf/17+/+f/7+Q1v//67G/bxv/0d/vzf8tvyT//2hvfNf/p3N//3v73hjT//95/dsI//42f4
         8//Ib8v/+Tc32P/nz27+M+u/yf7pjX+W/c+tf4r+5z+9+X/9JPr/+unN/9v6J9k/vvF/Z/8L6x+j/8WP
